[Modified] upload music 컨테이너 분리
